+++ modules/pim.nix
@@ -1,9 +1,10 @@
 {
   nixosModule = { user, pkgs, ... }: {
     environment.systemPackages = builtins.attrValues {
-      inherit (pkgs.gnome) gnome-contacts gnome-control-center;
+      inherit (pkgs.gnome) gnome-calendar gnome-contacts gnome-control-center;
     };
 
     services.gnome.gnome-online-accounts.enable = true;
+    services.gnome.evolution-data-server.enable = true;
   };
 }